Bug fixes: summary message, handling of comment lines. Added new field files.
Bug fix #1: End-of-sim message only showed -seed option even when -f option was used
Bug fix #2: A ":" in a comment line caused it to be processed as a valid configuration line. Added check for "#" as the first character of the line.
New field files for Period 15 oscillator and 3 known methuselahs (small but long-lived initial populations)